ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
kafka-python==2.0.2
better_profanity==0.7.0
pydantic==2.10.6
pytest==8.3.5
pytest-mock==3.14.0
httpx==0.28.1